🏗️ Move `useServerErrors.js` hook to new root `/src/hooks/` folder
In the new feature-based architecture, every hook that can be used by any feature have to be in the root `src/hooks/` folder, improving the discoverability of the app.
