+++ src/components/Forms/OrganizationForm/index.jsx
@@ -3,7 +3,7 @@ import { TextField, Button, CircularProgress } from '@mui/material'
 import { useForm } from 'react-hook-form'
 import { useNavigate } from 'react-router-dom'
 import { api } from '../../../services/api'
-import { useServerErrors } from '../useServerErrors'
+import { useServerErrors } from '../../../hooks/useServerErrors'
 
 import './style.scss'
